move no_other_refs
